Fix some issues + pre-final stretch
At this moment, this project is heading towards completion. Therefore, being delivered upon my completion of college.
I am close to completing what I proposed to be delivered. But this does not mean that it will be the end of this project, this game is just a fraction of the entire great project that is Knight of Brass